PackageInfo.g: Use proper UTF-8 accents in authors' names
This essentially reverts a commit from 2018 whose commit message said:

> Change accented letters in names
>
> This is to have them shown correctly on the web page.

But it is not clear to me which website this refers to? If it is
<https://gap-packages.github.io/permut> then I don't see how UTF-8
should be a problem (plenty packages use it for their author names).
However it may be, but the TeX-style accents previously used definitely
won't be rendered in a useful way